feat: unflag astro:assets
