+++ packages/astro/package.json
@@ -34,8 +34,6 @@
     "./env": "./env.d.ts",
     "./types": "./types.d.ts",
     "./client": "./client.d.ts",
-    "./client-base": "./client-base.d.ts",
-    "./client-image": "./client-image.d.ts",
     "./import-meta": "./import-meta.d.ts",
     "./astro-jsx": "./astro-jsx.d.ts",
     "./tsconfigs/*.json": "./tsconfigs/*",
@@ -91,8 +89,6 @@
     "zod.mjs",
     "env.d.ts",
     "client.d.ts",
-    "client-base.d.ts",
-    "client-image.d.ts",
     "content-types.template.d.ts",
     "content-module.template.mjs",
     "import-meta.d.ts",
