+++ packages/astro/src/core/config/config.ts
@@ -124,8 +124,6 @@ export function resolveFlags(flags: Partial<Flags>): CLIFlags {
 		host:
 			typeof flags.host === 'string' || typeof flags.host === 'boolean' ? flags.host : undefined,
 		drafts: typeof flags.drafts === 'boolean' ? flags.drafts : undefined,
-		experimentalAssets:
-			typeof flags.experimentalAssets === 'boolean' ? flags.experimentalAssets : undefined,
 	};
 }
