+++ packages/astro/src/core/config/schema.ts
@@ -44,7 +44,6 @@ const ASTRO_CONFIG_DEFAULTS = {
 	legacy: {},
 	redirects: {},
 	experimental: {
-		assets: false,
 		viewTransitions: false,
 		optimizeHoistedScript: false,
 	},
@@ -241,7 +240,6 @@ export const AstroConfigSchema = z.object({
 		.default(ASTRO_CONFIG_DEFAULTS.vite),
 	experimental: z
 		.object({
-			assets: z.boolean().optional().default(ASTRO_CONFIG_DEFAULTS.experimental.assets),
 			viewTransitions: z
 				.boolean()
 				.optional()
