+++ packages/astro/src/core/config/settings.ts
@@ -17,7 +17,6 @@ export function createBaseSettings(config: AstroConfig): AstroSettings {
 		config,
 		tsConfig: undefined,
 		tsConfigPath: undefined,
-
 		adapter: undefined,
 		injectedRoutes: [],
 		resolvedInjectedRoutes: [],
