+++ packages/astro/src/core/dev/container.ts
@@ -50,11 +50,7 @@ export async function createContainer({
 		isRestart,
 	});
 
-	// HACK: Since we only inject the endpoint if `experimental.assets` is on and it's possible for an integration to
-	// add that flag, we need to only check and inject the endpoint after running the config setup hook.
-	if (settings.config.experimental.assets) {
-		settings = injectImageEndpoint(settings);
-	}
+	settings = injectImageEndpoint(settings);
 
 	const { host, headers, open } = settings.config.server;
