+++ packages/astro/src/vite-plugin-inject-env-ts/index.ts
@@ -50,26 +50,6 @@ export async function setUpEnvTs({
 	if (fs.existsSync(envTsPath)) {
 		let typesEnvContents = await fs.promises.readFile(envTsPath, 'utf-8');
 
-		// TODO: Remove this logic in 3.0, as `astro/client-image` will be merged into `astro/client`
-		if (settings.config.experimental.assets && typesEnvContents.includes('types="astro/client"')) {
-			typesEnvContents = typesEnvContents.replace(
-				'types="astro/client"',
-				'types="astro/client-image"'
-			);
-			await fs.promises.writeFile(envTsPath, typesEnvContents, 'utf-8');
-			info(logging, 'assets', `Added ${bold(envTsPathRelativetoRoot)} types`);
-		} else if (
-			!settings.config.experimental.assets &&
-			typesEnvContents.includes('types="astro/client-image"')
-		) {
-			typesEnvContents = typesEnvContents.replace(
-				'types="astro/client-image"',
-				'types="astro/client"'
-			);
-			await fs.promises.writeFile(envTsPath, typesEnvContents, 'utf-8');
-			info(logging, 'assets', `Removed ${bold(envTsPathRelativetoRoot)} types`);
-		}
-
 		if (!fs.existsSync(dotAstroDir))
 			// Add `.astro` types reference if none exists
 			return;
@@ -83,11 +63,7 @@ export async function setUpEnvTs({
 	} else {
 		// Otherwise, inject the `env.d.ts` file
 		let referenceDefs: string[] = [];
-		if (settings.config.experimental.assets) {
-			referenceDefs.push('/// <reference types="astro/client-image" />');
-		} else {
-			referenceDefs.push('/// <reference types="astro/client" />');
-		}
+		referenceDefs.push('/// <reference types="astro/client" />');
 
 		if (fs.existsSync(dotAstroDir)) {
 			referenceDefs.push(dotAstroTypeReference);
