+++ packages/astro/test/core-image.test.js
@@ -20,9 +20,6 @@ describe('astro:image', () => {
 		before(async () => {
 			fixture = await loadFixture({
 				root: './fixtures/core-image/',
-				experimental: {
-					assets: true,
-				},
 				image: {
 					service: testImageService({ foo: 'bar' }),
 				},
@@ -434,9 +431,6 @@ describe('astro:image', () => {
 		before(async () => {
 			fixture = await loadFixture({
 				root: './fixtures/core-image-errors/',
-				experimental: {
-					assets: true,
-				},
 				image: {
 					service: testImageService(),
 				},
@@ -502,9 +496,6 @@ describe('astro:image', () => {
 		before(async () => {
 			fixture = await loadFixture({
 				root: './fixtures/core-image-base/',
-				experimental: {
-					assets: true,
-				},
 				image: {
 					service: testImageService(),
 				},
@@ -558,9 +549,6 @@ describe('astro:image', () => {
 				root: './fixtures/core-image-ssr/',
 				output: 'server',
 				adapter: testAdapter(),
-				experimental: {
-					assets: true,
-				},
 				image: {
 					service: testImageService(),
 				},
@@ -582,9 +570,6 @@ describe('astro:image', () => {
 		before(async () => {
 			fixture = await loadFixture({
 				root: './fixtures/core-image-ssg/',
-				experimental: {
-					assets: true,
-				},
 				image: {
 					service: testImageService(),
 				},
@@ -748,9 +733,6 @@ describe('astro:image', () => {
 				root: './fixtures/core-image-ssr/',
 				output: 'server',
 				adapter: testAdapter(),
-				experimental: {
-					assets: true,
-				},
 				image: {
 					service: testImageService(),
 				},
@@ -775,9 +757,6 @@ describe('astro:image', () => {
 				root: './fixtures/core-image-ssr/',
 				output: 'server',
 				adapter: testAdapter(),
-				experimental: {
-					assets: true,
-				},
 				image: {
 					service: testImageService(),
 				},
