+++ packages/astro/test/fixtures/content-collection-references/astro.config.mjs
@@ -3,9 +3,6 @@ import { testImageService } from '../../test-image-service.js';
 
 // https://astro.build/config
 export default defineConfig({
-	experimental: {
-		assets: true,
-	},
 	image: {
 		service: testImageService(),
 	},
