+++ packages/astro/test/units/content-collections/get-entry-type.test.js
@@ -28,14 +28,13 @@ const fixtures = [
 const contentFileExts = ['.md', '.mdx'];
 const dataFileExts = ['.yaml', '.yml', '.json'];
 
-// TODO: Remove `getEntryType` last parameter once `experimental.assets` is no longer experimental
 describe('Content Collections - getEntryType', () => {
 	fixtures.forEach(({ title, contentPaths }) => {
 		describe(title, () => {
 			it('Returns "content" for Markdown files', () => {
 				for (const entryPath of ['blog/first-post.md', 'blog/first-post.mdx']) {
 					const entry = fileURLToPath(new URL(entryPath, contentPaths.contentDir));
-					const type = getEntryType(entry, contentPaths, contentFileExts, dataFileExts, false);
+					const type = getEntryType(entry, contentPaths, contentFileExts, dataFileExts);
 					expect(type).to.equal('content');
 				}
 			});
@@ -47,7 +46,7 @@ describe('Content Collections - getEntryType', () => {
 					'banners/welcome.yml',
 				]) {
 					const entry = fileURLToPath(new URL(entryPath, contentPaths.contentDir));
-					const type = getEntryType(entry, contentPaths, contentFileExts, dataFileExts, false);
+					const type = getEntryType(entry, contentPaths, contentFileExts, dataFileExts);
 					expect(type).to.equal('data');
 				}
 			});
@@ -55,50 +54,50 @@ describe('Content Collections - getEntryType', () => {
 			it('Returns "content" for Markdown files in nested directories', () => {
 				for (const entryPath of ['blog/2021/01/01/index.md', 'blog/2021/01/01/index.mdx']) {
 					const entry = fileURLToPath(new URL(entryPath, contentPaths.contentDir));
-					const type = getEntryType(entry, contentPaths, contentFileExts, dataFileExts, false);
+					const type = getEntryType(entry, contentPaths, contentFileExts, dataFileExts);
 					expect(type).to.equal('content');
 				}
 			});
 
 			it('Returns "config" for config files', () => {
 				const entry = fileURLToPath(contentPaths.config.url);
-				const type = getEntryType(entry, contentPaths, contentFileExts, dataFileExts, false);
+				const type = getEntryType(entry, contentPaths, contentFileExts, dataFileExts);
 				expect(type).to.equal('config');
 			});
 
 			it('Returns "unsupported" for non-Markdown files', () => {
 				const entry = fileURLToPath(new URL('blog/robots.txt', contentPaths.contentDir));
-				const type = getEntryType(entry, contentPaths, contentFileExts, dataFileExts, false);
+				const type = getEntryType(entry, contentPaths, contentFileExts, dataFileExts);
 				expect(type).to.equal('unsupported');
 			});
 
 			it('Returns "ignored" for .DS_Store', () => {
 				const entry = fileURLToPath(new URL('blog/.DS_Store', contentPaths.contentDir));
-				const type = getEntryType(entry, contentPaths, contentFileExts, dataFileExts, false);
+				const type = getEntryType(entry, contentPaths, contentFileExts, dataFileExts);
 				expect(type).to.equal('ignored');
 			});
 
 			it('Returns "ignored" for unsupported files using an underscore', () => {
 				const entry = fileURLToPath(new URL('blog/_draft-robots.txt', contentPaths.contentDir));
-				const type = getEntryType(entry, contentPaths, contentFileExts, dataFileExts, false);
+				const type = getEntryType(entry, contentPaths, contentFileExts, dataFileExts);
 				expect(type).to.equal('ignored');
 			});
 
 			it('Returns "ignored" when using underscore on file name', () => {
 				const entry = fileURLToPath(new URL('blog/_first-post.md', contentPaths.contentDir));
-				const type = getEntryType(entry, contentPaths, contentFileExts, dataFileExts, false);
+				const type = getEntryType(entry, contentPaths, contentFileExts, dataFileExts);
 				expect(type).to.equal('ignored');
 			});
 
 			it('Returns "ignored" when using underscore on directory name', () => {
 				const entry = fileURLToPath(new URL('blog/_draft/first-post.md', contentPaths.contentDir));
-				const type = getEntryType(entry, contentPaths, contentFileExts, dataFileExts, false);
+				const type = getEntryType(entry, contentPaths, contentFileExts, dataFileExts);
 				expect(type).to.equal('ignored');
 			});
 
 			it('Returns "ignored" for images', () => {
 				const entry = fileURLToPath(new URL('blog/first-post.png', contentPaths.contentDir));
-				const type = getEntryType(entry, contentPaths, contentFileExts, dataFileExts, true);
+				const type = getEntryType(entry, contentPaths, contentFileExts, dataFileExts);
 				expect(type).to.equal('ignored');
 			});
 		});
