+++ packages/integrations/markdoc/src/content-entry-type.ts
@@ -1,6 +1,7 @@
 import type { Config as MarkdocConfig, Node } from '@markdoc/markdoc';
 import Markdoc from '@markdoc/markdoc';
 import type { AstroConfig, ContentEntryType } from 'astro';
+import { emitESMImage } from 'astro/assets/utils';
 import matter from 'gray-matter';
 import fs from 'node:fs';
 import path from 'node:path';
@@ -8,13 +9,12 @@ import { fileURLToPath } from 'node:url';
 import type * as rollup from 'rollup';
 import type { ErrorPayload as ViteErrorPayload } from 'vite';
 import type { ComponentConfig } from './config.js';
-import { MarkdocError, isComponentConfig, isValidUrl, prependForwardSlash } from './utils.js';
-import { emitESMImage } from 'astro/assets/utils';
 import { htmlTokenTransform } from './html/transform/html-token-transform.js';
 import type { MarkdocConfigResult } from './load-config.js';
 import type { MarkdocIntegrationOptions } from './options.js';
 import { setupConfig } from './runtime.js';
 import { getMarkdocTokenizer } from './tokenizer.js';
+import { MarkdocError, isComponentConfig, isValidUrl, prependForwardSlash } from './utils.js';
 
 export async function getContentEntryType({
 	markdocConfigResult,
@@ -96,13 +96,11 @@ export async function getContentEntryType({
 				});
 			}
 
-			if (astroConfig.experimental.assets) {
-				await emitOptimizedImages(ast.children, {
-					astroConfig,
-					pluginContext,
-					filePath,
-				});
-			}
+			await emitOptimizedImages(ast.children, {
+				astroConfig,
+				pluginContext,
+				filePath,
+			});
 
 			const res = `import { Renderer } from '@astrojs/markdoc/components';
 import { createGetHeadings, createContentComponent } from '@astrojs/markdoc/runtime';
@@ -110,12 +108,10 @@ ${
 	markdocConfigUrl
 		? `import markdocConfig from ${JSON.stringify(markdocConfigUrl.pathname)};`
 		: 'const markdocConfig = {};'
-}${
-				astroConfig.experimental.assets
-					? `\nimport { experimentalAssetsConfig } from '@astrojs/markdoc/experimental-assets-config';
-markdocConfig.nodes = { ...experimentalAssetsConfig.nodes, ...markdocConfig.nodes };`
-					: ''
-			}
+}
+
+import { assetsConfig } from '@astrojs/markdoc/runtime-assets-config';
+markdocConfig.nodes = { ...assetsConfig.nodes, ...markdocConfig.nodes };
 
 ${getStringifiedImports(componentConfigByTagMap, 'Tag', astroConfig.root)}
 ${getStringifiedImports(componentConfigByNodeMap, 'Node', astroConfig.root)}
