+++ packages/integrations/markdoc/src/runtime.ts
@@ -7,11 +7,10 @@ import Markdoc, {
 } from '@markdoc/markdoc';
 import type { AstroInstance } from 'astro';
 import { createComponent, renderComponent } from 'astro/runtime/server/index.js';
-import type { AstroMarkdocConfig } from './config.js';
+import { type AstroMarkdocConfig } from './config.js';
 import { setupHeadingConfig } from './heading-ids.js';
 import { htmlTag } from './html/tagdefs/html.tag.js';
 import type { MarkdocIntegrationOptions } from './options.js';
-
 /**
  * Merge user config with default config and set up context (ex. heading ID slugger)
  * Called on each file's individual transform.
