+++ packages/integrations/markdoc/test/fixtures/image-assets/astro.config.mjs
@@ -1,12 +1,9 @@
-import { defineConfig } from 'astro/config';
 import markdoc from '@astrojs/markdoc';
+import { defineConfig } from 'astro/config';
 import { testImageService } from '../../../../../astro/test/test-image-service.js';
 
 // https://astro.build/config
 export default defineConfig({
-	experimental: {
-		assets: true,
-	},
 	image: {
 		service: testImageService(),
 	},
