+++ packages/integrations/netlify/test/hosted/hosted-astro-project/astro.config.mjs
@@ -5,7 +5,4 @@ import { defineConfig } from 'astro/config';
 export default defineConfig({
 	output: 'server',
 	adapter: netlify(),
-	experimental: {
-		assets: true,
-	},
 });
