+++ packages/integrations/vercel/test/edge-middleware.test.js
@@ -10,7 +10,7 @@ describe('Vercel edge middleware', () => {
 		await fixture.build();
 		const contents = await fixture.readFile(
 			// this is abysmal...
-			'../.vercel/output/functions/render.func/packages/integrations/vercel/test/fixtures/middleware-with-edge-file/dist/middleware.mjs'
+			'../.vercel/output/functions/render.func/www/withastro/astro/packages/integrations/vercel/test/fixtures/middleware-with-edge-file/dist/middleware.mjs'
 		);
 		expect(contents.includes('title:')).to.be.true;
 		chaiJestSnapshot.setTestName('Middleware with handler file');
@@ -24,7 +24,7 @@ describe('Vercel edge middleware', () => {
 		await fixture.build();
 		const contents = await fixture.readFile(
 			// this is abysmal...
-			'../.vercel/output/functions/render.func/packages/integrations/vercel/test/fixtures/middleware-without-edge-file/dist/middleware.mjs'
+			'../.vercel/output/functions/render.func/www/withastro/astro/packages/integrations/vercel/test/fixtures/middleware-without-edge-file/dist/middleware.mjs'
 		);
 		expect(contents.includes('title:')).to.be.false;
 		chaiJestSnapshot.setTestName('Middleware without handler file');
