+++ packages/integrations/vercel/test/hosted/hosted-astro-project/astro.config.mjs
@@ -5,7 +5,4 @@ import { defineConfig } from 'astro/config';
 export default defineConfig({
 	output: 'server',
 	adapter: vercel(),
-	experimental: {
-		assets: true,
-	},
 });
